Change soil requirement for mutations to a predicate
Since we can't rely on testing for farmland, for instance, if we want to also allow mutations to spread to modded soils, we need a predicate instead. Also, I added a mutation recipe for sugar cane - it worked out of the box which was sweet, although you need to remember to have water next to your sand.